+++ src/parser/AzurePipelineDefinitions.ts
@@ -8,10 +8,6 @@ export interface AzurePipelineTaskDefiniton {
 // Definition from https://github.com/microsoft/azure-pipelines-task-lib/blob/master/tasks.schema.json
 export interface AzurePipelineTaskParameterDefiniton {
   name: string;
-  type: "string" | "boolean" | "filePath";
-  label: string;
   required?: boolean;
   defaultValue?: string | boolean;
-  visibleRule?: string;
-  helpMarkDown?: string;
 }
